Add files via upload
The name of the people in my group is Rosairi and Chris. We first started getting the hang go R studio and then started looking at the data how much data rstudio can hold. we found it interesting how they included most important factors one that stuck out to me was the LGBTQ.